The version of port forwarding stuff which passed Cedar(ReliSock) level testing. Major features are: 	- Forwarding server: Fault tolerant by 		: periodically synchronizing Kernel, Condor_Persistent File, 		  and in-memory Forwarding rules 		: periodically checking heartbeat of Cedars which have 		  forwarding being set 	- Cedar: I claim it is efficient because 		: it sets up forwarding rule after binding a port 		: it unsets forwarding rule before actively connecting 		  to the peer 		: it directly connects to the peer which is in the same 		  private network

+++ src/condor_io/reli_sock.C
@@ -428,7 +428,7 @@ ReliSock::greetMnger()
     }
 
     // Bind _mngSock
-    int lowPort, highPort;
+    unsigned short lowPort, highPort;
     if ( get_port_range(&lowPort, &highPort) == TRUE ) {
         if ( bindWithin(_mngSock, lowPort, highPort) != TRUE ) {
             return FALSE;
@@ -447,7 +447,7 @@ ReliSock::greetMnger()
 
     // get (ip-addr, port) of netMnger
 	char mngerHost[50];
-	int mngerPort;
+	unsigned short mngerPort;
 	if ( getMnger (mngerHost, &mngerPort) != TRUE ) {
 		dprintf (D_ALWAYS, "ReliSock::greetMnger -\
 							failed to get NetMnger info\n");
@@ -457,7 +457,7 @@ ReliSock::greetMnger()
     // Connect to netMnger
     bzero(&sockAddr, sizeof(sockAddr));
     sockAddr.sin_family = AF_INET;
-    unsigned long mngerIP;
+    unsigned int mngerIP;
     if((mngerIP = inet_addr(mngerHost)) != (unsigned)-1) {
         memcpy(&sockAddr.sin_addr, &mngerIP, sizeof(mngerIP));
     } else {
